Fixed query bug and Added activityLogBean
findRecentLog query was missing ORDER BY
-activityLogBean created to initialized bean to show updated log into the front end

+++ src/main/java/com/sil/asset_tagging_system/model/ActivityLog.java
@@ -43,7 +43,7 @@ public class ActivityLog {
     private UUID correlationId;
 
     @Column(name="sequence_in_action")
-    private Byte sequenceInAction;
+    private Short sequenceInAction;
 
     @Column(name="actor_user_id")
     private Long actorUserId;
